+++ chatlab/builtins/__init__.py
@@ -1,0 +1,5 @@
+"""Builtins for ChatLab."""
+
+from .python import run_cell
+
+__all__ = ["run_cell"]
